# api/swagger/docs/swagger.yaml
basePath: /
definitions:
  common.SwaggerErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      data: {}
      message:
        example: invalid_request
        type: string
    type: object
  common.SwaggerResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: ok
        type: string
    type: object
...
  v1.CreateDocumentRequest:
    properties:
      category_id:
        type: integer
      confidence:
        type: number
      content:
        type: string
      cover_url:
        type: string
      slug:
        type: string
      source:
        type: string
      status:
        type: string
      summary:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        type: string
    type: object
...
  v1.DocumentCategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      slug:
        type: string
    type: object
  v1.DocumentResponse:
    properties:
      author_id:
        type: integer
      category:
        $ref: '#/definitions/v1.DocumentCategoryResponse'
      category_id:
        type: integer
      confidence:
        type: number
      content:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      published_at:
        type: string
      slug:
        type: string
      source:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          $ref: '#/definitions/v1.DocumentTagResponse'
        type: array
      title:
        type: string
      updated_at:
        type: string
      word_count:
        type: integer
    type: object
  v1.DocumentTagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  v1.ListDocumentsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.DocumentResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  v1.UpdateDocumentRequest:
    properties:
      category_id:
        type: integer
      confidence:
        type: number
      content:
        type: string
      cover_url:
        type: string
      slug:
        type: string
      source:
        type: string
      status:
        type: string
      summary:
        type: string
      tag_ids:
        items:
          type: integer
        type: array
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: 'REST API for Knowledge Core (v1). JSON envelope: code (HTTP-style),
    message, data. 中文：Knowledge Core REST API（v1）；响应为 code、message、data 的 JSON 包装。'
  title: Knowledge Core HTTP API
  version: "1.0"
paths:
...
  /api/v1/admin/documents:
    get:
      description: List all documents for admin management.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Keyword
        in: query
        name: q
        type: string
      - description: Category slug or path
        in: query
        name: category
        type: string
      - description: Tag slug or name
        in: query
        name: tag
        type: string
      - description: Status
        enum:
        - draft
        - published
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.ListDocumentsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List documents
      tags:
      - Admin Documents
    post:
      consumes:
      - application/json
      description: Create a Markdown document and index metadata. Admin only.
      parameters:
      - description: Document create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.DocumentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create document
      tags:
      - Admin Documents
  /api/v1/admin/documents/{id}:
    delete:
      description: Delete document metadata and Markdown file. Admin only.
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SwaggerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - Admin Documents
    get:
      description: Get a document by id. Admin only.
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.DocumentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get document
      tags:
      - Admin Documents
    patch:
      consumes:
      - application/json
      description: Update document metadata, content, tags, category, or status. Admin
        only.
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: Document update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.DocumentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update document
      tags:
      - Admin Documents
...
  /api/v1/documents:
    get:
      description: List published documents with optional keyword, category, and tag
        filters.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Keyword
        in: query
        name: q
        type: string
      - description: Category slug or path
        in: query
        name: category
        type: string
      - description: Tag slug or name
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.ListDocumentsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      summary: List published documents
      tags:
      - Documents
  /api/v1/documents/{id}:
    get:
      description: Get a published document by id.
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.SwaggerResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.DocumentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.SwaggerErrorResponse'
      summary: Get published document
      tags:
      - Documents
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Use the format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"